# docs/API_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: Brix Nutritional App API
  version: 2.0.0
  description: |
    REST API для Brix Nutritional App (Mobile + Admin Web Panel).
    
    ## Компоненты системы
    
    ### 1. Mobile App (Flutter)
    Мобильное приложение для пользователей (iOS/Android).
    Использует все endpoints из категорий: Auth, Users, Home, Meal Plans, Recipes, Diary, AI Chat, Lab Tests, Courses, Blog, Notifications, Subscriptions.
    
    ### 2. Admin Web Panel (Next.js)
    Веб-панель для администраторов и контент-менеджеров.
    Использует admin-специфичные endpoints и endpoints для управления контентом (Courses, Lessons, Recipes, Meal Plans, Users, Blog, Subscriptions, Analytics).
    
    ### 3. Backend API (Strapi)
    Сервер на Node.js + Strapi CMS с кастомными контроллерами.
    
    ---
    
    ## Авторизация
    
    ### Для Mobile App:
    Большинство endpoints требуют JWT токен в заголовке:
    ```
    Authorization: Bearer <token>
    ```
    
    ### Для Admin Web Panel:
    Admin endpoints требуют Strapi admin JWT токен:
    ```
    Authorization: Bearer <admin-token>
    ```
    Токен получается через Strapi `/admin/login` endpoint.
    
    ---
    
    ## Rate Limiting
    - Auth endpoints: 5 запросов / 15 минут
    - Mobile App endpoints: 100 запросов / минуту
    - Admin endpoints: 1000 запросов / минуту
    
  contact:
    name: API Support
    email: support@brix-nutrition.com
servers:
  - url: http://localhost:1337/api
    description: Development
...
paths:
...
  /meal-plan/current:
    get:
      tags: [Meal Plans]
      summary: Получить текущий план питания
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Текущий план
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MealPlan'
  /meal-plan/day/{date}:
    get:
      tags: [Meal Plans]
      summary: Получить план на конкретный день
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
            example: "2025-10-10"
      responses:
        '200':
          description: План на день
          content:
            application/json:
              schema:
                type: object
                properties:
                  date:
                    type: string
                    format: date
                  meals:
                    type: array
                    items:
                      $ref: '#/components/schemas/MealSlot'
  /meal-plan/replace:
    post:
      tags: [Meal Plans]
      summary: Заменить блюдо в плане
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mealSlotId, newRecipeId]
              properties:
                mealSlotId:
                  type: string
                  format: uuid
                newRecipeId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Блюдо заменено
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  updatedMeal:
                    $ref: '#/components/schemas/MealSlot'

  # ============= RECIPES =============
...
  /diary/day/{date}:
    get:
      tags: [Diary]
      summary: Получить дневник за день
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Дневник
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiaryDay'
...
  /diary/water:
    post:
      tags: [Diary]
      summary: Обновить количество воды
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [date, increment]
              properties:
                date:
                  type: string
                  format: date
                increment:
                  type: integer
                  description: Изменение в мл (может быть отрицательным)
                  example: 200
      responses:
        '200':
          description: Вода обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaterLog'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MealPlan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        type:
          type: string
        startDate:
          type: string
          format: date
        meals:
          type: array
          items:
            $ref: '#/components/schemas/MealSlot'
    MealSlot:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [wakeup, breakfast, snack, lunch, dinner, sleep]
        time:
          type: string
          example: "08:00"
        recipe:
          $ref: '#/components/schemas/Recipe'
        portion:
          type: integer
          description: Граммы
        calories:
          type: integer
        importance:
          type: string
          description: Описание важности приема пищи
    Recipe:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        prepTime:
          type: integer
          description: Минуты
        calories:
          type: integer
        protein:
          type: number
        carbs:
          type: number
        fats:
          type: number
        instructions:
          type: array
          items:
            type: string
        ingredients:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              amount:
                type: number
              unit:
                type: string
        tags:
          type: array
          items:
            type: string
    DiaryDay:
      type: object
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        meals:
          type: array
          items:
            $ref: '#/components/schemas/MealEntry'
        waterLog:
          $ref: '#/components/schemas/WaterLog'
        moodRating:
          type: integer
          minimum: 1
          maximum: 5
        isCompleted:
          type: boolean
    MealEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mealName:
          type: string
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        consumedAt:
          type: string
          format: date-time
        portionGrams:
          type: integer
        calories:
          type: integer
        photoUrl:
          type: string
        fromPlan:
          type: boolean
        recipeId:
          type: string
          format: uuid
    WaterLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        totalAmount:
          type: integer
          description: Миллилитры
        goal:
          type: integer
          description: Целевое количество мл
...